fixed nibscan to display halftracks properly, now that they are supported.  youll get mismatches comparing an image with HT verses one without (due to the empty halftracks)
git-svn-id: https://c64preservation.com/svn/nibtools/trunk@20 048eaad1-c872-4173-8c30-523bb29f53ec

+++ nibscan.c
@@ -271,11 +271,11 @@ compare_disks(void)
 		track_density[track] %= BM_MATCH;
 		track_density2[track] %= BM_MATCH;
 
-		printf("Track %2d, Disk 1: (%d) %d\n",
-		  track / 2, (track_density[track] & 3), track_length[track]);
+		printf("Track %4.1f, Disk 1: (%d) %d\n",
+		 	(float) track / 2, (track_density[track] & 3), track_length[track]);
 
-		printf("Track %2d, Disk 2: (%d) %d\n",
-		  track / 2, (track_density2[track] & 3), track_length2[track]);
+		printf("Track %4.1f, Disk 2: (%d) %d\n",
+		 	(float) track / 2, (track_density2[track] & 3), track_length2[track]);
 
 		if( ((track_length[track] > 0) && (track_length2[track] == 0)) ||
 			((track_length[track] == 0) && (track_length2[track] > 0)) )
